# Improve performance of Buf::get_*()