@carllerche 2018-04-27
I released this to crates.io: https://crates.io/crates/bytes/0.4.7.